Rework reminders code
