+++ packages/server/customer-os-api-sdk/graph/resolver/reminders.resolvers.go
@@ -14,17 +14,17 @@ import (
 
 // ReminderCreate is the resolver for the reminder_Create field.
 func (r *mutationResolver) ReminderCreate(ctx context.Context, input model.ReminderInput) (*model.Reminder, error) {
-	panic(fmt.Errorf("not implemented: ReminderCreate - reminder_Create"))
+	panic(fmt.Errorf("not implemented: CreateReminder - reminder_Create"))
 }
 
 // ReminderUpdate is the resolver for the reminder_Update field.
 func (r *mutationResolver) ReminderUpdate(ctx context.Context, input model.ReminderUpdateInput) (*model.Reminder, error) {
-	panic(fmt.Errorf("not implemented: ReminderUpdate - reminder_Update"))
+	panic(fmt.Errorf("not implemented: UpdateReminder - reminder_Update"))
 }
 
 // Reminder is the resolver for the reminder field.
 func (r *queryResolver) Reminder(ctx context.Context, id string) (*model.Reminder, error) {
-	panic(fmt.Errorf("not implemented: Reminder - reminder"))
+	panic(fmt.Errorf("not implemented: GetReminderById - reminder"))
 }
 
 // RemindersForOrganization is the resolver for the remindersForOrganization field.
