+++ packages/server/events-processing-platform/.gitignore
@@ -1,3 +1,3 @@
 bin
-events-processing-platform
+events-processing-platform-subscribers
 tmp-sources
