--- packages/server/events-processing-platform/domain/reminder/events/events.go
@@ -1,13 +1,0 @@
-package events
-
-const (
-	ReminderCreateV1       = "V1_REMINDER_CREATE"
-	ReminderUpdateV1       = "V1_REMINDER_UPDATE"
-	ReminderNotificationV1 = "V1_REMINDER_NOTIFICATION"
-)
-
-const (
-	FieldMaskContent   = "content"
-	FieldMaskDueDate   = "dueDate"
-	FieldMaskDismissed = "dismissed"
-)
