+++ packages/server/events-processing-platform/service/test/reminder_service_it_test.go
@@ -5,8 +5,7 @@ import (
 	"testing"
 
 	"github.com/openline-ai/openline-customer-os/packages/server/customer-os-common-module/utils"
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/domain/reminder/aggregate"
-	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/domain/reminder/events"
+	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/domain/reminder"
 	"github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/test"
 	eventstoret "github.com/openline-ai/openline-customer-os/packages/server/events-processing-platform/test/eventstore"
 	commonpb "github.com/openline-ai/openline-customer-os/packages/server/events-processing-proto/gen/proto/go/api/grpc/v1/common"
@@ -29,7 +28,7 @@ func TestReminderService_CreateReminder(t *testing.T) {
 	response, err := reminderClient.CreateReminder(ctx, &reminderpb.CreateReminderGrpcRequest{
 		Tenant:         tenant,
 		Content:        "New Reminder",
-		UserId:         "user",
+		LoggedInUserId: "user",
 		OrganizationId: "org",
 		Dismissed:      false,
 		DueDate:        utils.ConvertTimeToTimestampPtr(utils.NowPtr()),
@@ -46,13 +45,13 @@ func TestReminderService_CreateReminder(t *testing.T) {
 	eventsMap := aggregateStore.GetEventMap()
 	require.Equal(t, 1, len(eventsMap))
 
-	reminderAggregate := aggregate.NewReminderAggregateWithTenantAndID(tenant, response.Id)
+	reminderAggregate := reminder.NewReminderAggregateWithTenantAndID(tenant, response.Id)
 	eventList := eventsMap[reminderAggregate.ID]
 	require.Equal(t, 1, len(eventList))
-	require.Equal(t, events.ReminderCreateV1, eventList[0].GetEventType())
-	require.Equal(t, string(aggregate.ReminderAggregateType)+"-"+tenant+"-"+reminderId, eventList[0].GetAggregateID())
+	require.Equal(t, reminder.ReminderCreateV1, eventList[0].GetEventType())
+	require.Equal(t, string(reminder.ReminderAggregateType)+"-"+tenant+"-"+reminderId, eventList[0].GetAggregateID())
 
-	var eventData events.ReminderCreateEvent
+	var eventData reminder.ReminderCreateEvent
 
 	err = eventList[0].GetJsonData(&eventData)
 	require.Nil(t, err, "Failed to unmarshal event data")
@@ -77,7 +76,7 @@ func TestReminderService_UpdateReminder(t *testing.T) {
 	createResponse, err := reminderClient.CreateReminder(ctx, &reminderpb.CreateReminderGrpcRequest{
 		Tenant:         tenant,
 		Content:        "New Reminder",
-		UserId:         "user",
+		LoggedInUserId: "user",
 		OrganizationId: "org",
 		Dismissed:      false,
 		DueDate:        utils.ConvertTimeToTimestampPtr(utils.NowPtr()),
@@ -93,11 +92,11 @@ func TestReminderService_UpdateReminder(t *testing.T) {
 	reminderId := createResponse.Id
 	eventsMap := aggregateStore.GetEventMap()
 	require.Equal(t, 1, len(eventsMap))
-	reminderAggregate := aggregate.NewReminderAggregateWithTenantAndID(tenant, createResponse.Id)
+	reminderAggregate := reminder.NewReminderAggregateWithTenantAndID(tenant, createResponse.Id)
 	eventList := eventsMap[reminderAggregate.ID]
 	require.Equal(t, 1, len(eventList))
-	require.Equal(t, events.ReminderCreateV1, eventList[0].GetEventType())
-	require.Equal(t, string(aggregate.ReminderAggregateType)+"-"+tenant+"-"+reminderId, eventList[0].GetAggregateID())
+	require.Equal(t, reminder.ReminderCreateV1, eventList[0].GetEventType())
+	require.Equal(t, string(reminder.ReminderAggregateType)+"-"+tenant+"-"+reminderId, eventList[0].GetAggregateID())
 
 	updateResponse, err := reminderClient.UpdateReminder(ctx, &reminderpb.UpdateReminderGrpcRequest{
 		Tenant:         tenant,
@@ -117,10 +116,10 @@ func TestReminderService_UpdateReminder(t *testing.T) {
 	require.Equal(t, 1, len(eventsMap))
 	eventList = eventsMap[reminderAggregate.ID]
 	require.Equal(t, 2, len(eventList))
-	require.Equal(t, events.ReminderUpdateV1, eventList[1].GetEventType())
-	require.Equal(t, string(aggregate.ReminderAggregateType)+"-"+tenant+"-"+reminderId, eventList[1].GetAggregateID())
+	require.Equal(t, reminder.ReminderUpdateV1, eventList[1].GetEventType())
+	require.Equal(t, string(reminder.ReminderAggregateType)+"-"+tenant+"-"+reminderId, eventList[1].GetAggregateID())
 
-	var eventData events.ReminderUpdateEvent
+	var eventData reminder.ReminderUpdateEvent
 
 	err = eventList[1].GetJsonData(&eventData)
 	require.Nil(t, err, "Failed to unmarshal event data")
