+++ packages/server/events-processing-proto/grpc/v1/reminder.proto
@@ -6,15 +6,14 @@ import "common/source.proto";
 option go_package = "api/grpc/v1/reminder;reminder_grpc_service";
 
 message CreateReminderGrpcRequest {
-    SourceFields sourceFields = 1;
-    google.protobuf.Timestamp createdAt = 2;
-    string content = 3;
-    google.protobuf.Timestamp dueDate = 4;
-    string userId = 5;
-    string organizationId = 6;
-    bool dismissed = 7;
-    string tenant = 8;
-    string loggedInUserId = 9;
+    string tenant = 1;
+    string loggedInUserId = 2;
+    SourceFields sourceFields = 3;
+    google.protobuf.Timestamp createdAt = 4;
+    string content = 5;
+    google.protobuf.Timestamp dueDate =6;
+    string organizationId = 7;
+    bool dismissed = 8;
 }
 
 enum ReminderFieldMask {
@@ -25,15 +24,16 @@ enum ReminderFieldMask {
 }
 
 message UpdateReminderGrpcRequest {
-    string app_source = 1;
-    string content = 2;
-    google.protobuf.Timestamp dueDate = 3;
-    bool dismissed = 4;
-    google.protobuf.Timestamp updatedAt = 5;
-    string reminderId = 6;
-    string tenant = 7;
-    repeated ReminderFieldMask fieldsMask = 8;
-    string loggedInUserId = 9;
+    string tenant = 1;
+    string loggedInUserId = 2;
+    string reminderId = 3;
+    string app_source = 4;
+    string content = 5;
+    google.protobuf.Timestamp dueDate = 6;
+    bool dismissed = 7;
+    google.protobuf.Timestamp updatedAt = 8;
+
+    repeated ReminderFieldMask fieldsMask = 9;
 }
 
 message ReminderGrpcResponse {
